+++ platforms/nextjs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @shallot-ui/platform-nextjs
 
+## 2.0.0-beta.5
+
+### Patch Changes
+
+- Theme mode variants extend existing varians
+- Updated dependencies
+  - @shallot-ui/platform-react@2.0.0-beta.5
+  - @shallot-ui/core-props@2.0.0-beta.5
+  - @shallot-ui/core-theme@2.0.0-beta.5
+
 ## 2.0.0-beta.4
 
 ### Patch Changes