+++ platforms/react-native/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @shallot-ui/platform-react-native
 
+## 2.0.0-beta.5
+
+### Patch Changes
+
+- Theme mode variants extend existing varians
+- Updated dependencies
+  - @shallot-ui/component-checkbox@3.0.0-beta.5
+  - @shallot-ui/component-button@3.0.0-beta.5
+  - @shallot-ui/component-switch@3.0.0-beta.5
+  - @shallot-ui/component-input@3.0.0-beta.5
+  - @shallot-ui/component-tag@3.0.0-beta.5
+  - @shallot-ui/core-mixins@2.0.0-beta.5
+  - @shallot-ui/core-props@2.0.0-beta.5
+  - @shallot-ui/core-theme@2.0.0-beta.5
+  - @shallot-ui/core-utils@2.0.0-beta.5
+
 ## 2.0.0-beta.4
 
 ### Patch Changes